requirements.txt
rich
python-okx
web3
httpx
retry
pyyaml
beaupy